fix typo in int64_t definition

+++ include/stdint.h
@@ -10,7 +10,7 @@ typedef unsigned short uint16_t;
 typedef signed long   int32_t;
 typedef unsigned long uint32_t;
 
-typedef long long int16_t;
+typedef long long int64_t;
 typedef unsigned long long uint64_t;
 
 typedef unsigned long uintptr_t;
